requirements.txt
imgui-bundle
glfw
pywinpty
pyte
huggingface_hub
Pillow
numpy
